feat(marketplace): replace two-page navigation with split-panel master-detail layout
Merges BrowsePage and PluginDetailPage into a single MarketplacePage
component. The list (40%) and detail (60%) panels are visible
simultaneously — selecting a plugin updates the detail inline instead
of navigating away. Selection state is bidirectionally synced with the
URL so direct links like /marketplace/research work as deep-links.

- Extract shared TrustBadge/TypeBadge into PluginBadges.tsx
- Add .row-list-item.selected CSS with inset accent bar
- Replace two routes with single optional-param route (marketplace/:slug?)
- Tag clicks in detail panel filter the list inline (no navigation)
- Related plugin clicks update selectedSlug within the same view
- Rewrite tests as MarketplacePage.test.tsx; delete old test files
